+++ app/src/main/java/com/kondenko/yamblzweather/ui/about/FragmentAbout.java
@@ -1,4 +1,4 @@
-package com.kondenko.yamblzweather.screens.about;
+package com.kondenko.yamblzweather.ui.about;
 
 
 import android.os.Bundle;
