+++ app/src/main/res/layout/fragment_about.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.kondenko.yamblzweather.screens.weather.FragmentWeather">
+    tools:context="com.kondenko.yamblzweather.ui.weather.FragmentWeather">
 
     <TextView
         style="@style/TextAppearance.AppCompat.Headline"
